refactor(desktop): move conversation markdown helper
