+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -10,6 +10,7 @@ const sourcePaths = [
   'app.tsx',
   'app-shell.tsx',
   'cached-theme-bootstrap.ts',
+  'conversation-markdown.ts',
 ] as const;
 
 export const RENDERER_SHELL_SOURCE_REPO_PATHS: readonly string[] = sourcePaths.map(
